+++ platform/src/codatplatform/companies.py
@@ -163,7 +163,7 @@ class Companies:
     
     def list(self, request: operations.ListCompaniesRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListCompaniesResponse:
         r"""List companies
-        Returns a list of your companies. The company schema contains a list of [connections](https://docs.codat.io/codat-api#/schemas/Connection) related to the company.
+        Returns a list of your companies. The company schema contains a list of [connections](https://docs.codat.io/platform-api#/schemas/Connection) related to the company.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())