+++ platform/src/codatplatform/models/operations/list_rules.py
@@ -27,10 +27,13 @@ class ListRulesRequest:
 @dataclasses.dataclass
 class ListRulesResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     webhooks: Optional[shared_webhooks.Webhooks] = dataclasses.field(default=None)
     r"""OK"""