+++ platform/src/codatplatform/models/shared/validationitem.py
@@ -12,7 +12,10 @@ from typing import Optional
 @dataclasses.dataclass
 class ValidationItem:
     item_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('itemId'), 'exclude': lambda f: f is None }})
+    r"""Unique identifier for a validation item."""
     message: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('message'), 'exclude': lambda f: f is None }})
+    r"""A message outlining validation item's issue."""
     validator_name: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('validatorName'), 'exclude': lambda f: f is None }})
+    r"""Name of validator."""